+++ app/components/SongList/components/AddSongBottomSheet.js
@@ -24,8 +24,9 @@ import * as StyleConstants from '@StyleConstants'
 import Song from 'Gruvee/lib/Song'
 import SongComment from 'Gruvee/lib/SongComment'
 
-const screenHeight = Dimensions.get('screen').height
-const navBarHeight = Platform.OS === 'ios' ? 80 : 84
+const windowHeight = Dimensions.get('window').height
+// These hardcoded values suck, but there isn't a solid way I can see to get the proper nav heights
+const screenHeight = Platform.OS === 'ios' ? windowHeight - 94 : windowHeight - 84
 const timesIcon = require('Gruvee/assets/icons/times/times_icon.png')
 
 // Styles
@@ -84,7 +85,7 @@ const AddSongBottomSheet = ({
     return (
         <BottomSheet
             ref={bottomSheetRef}
-            snapPoints={[screenHeight - navBarHeight, screenHeight / 2 + navBarHeight, 0]}
+            snapPoints={[screenHeight, screenHeight / 2, 0]}
             borderRadius={10}
             initialSnap={2}
             renderContent={generateSheetContent(
