+++ app/config/navigation/constants.js
@@ -1,0 +1,26 @@
+// Navigation IDs
+export const STACK_ID = 'navigation.gruvee.stack.id'
+
+export const ROOT_NAV_ID = 'navigation.gruvee.app.id'
+export const ROOT_NAV_NAME = 'navigation.gruvee.app'
+
+export const AUTH_NAV_ID = 'navigation.gruvee.auth.id'
+export const AUTH_NAV_NAME = 'Auth'
+
+export const PLAYLIST_NAV_ID = 'navigation.gruvee.playlist.id'
+export const PLAYLIST_NAV_NAME = 'Playlists'
+
+export const SONG_LIST_NAV_ID = 'navigation.gruvee.songList.id'
+export const SONG_LIST_NAV_NAME = 'Songs'
+
+export const COMMENTS_LIST_NAV_ID = 'navigation.gruvee.commentsList.id'
+export const COMMENTS_LIST_NAV_NAME = 'navigation.gruvee.commentsList'
+
+export const TOP_BAR_MEMBERS_ACTION_ID = 'navigation.gruvee.showMembersAction.id'
+export const TOP_BAR_MEMBERS_ACTION_NAME = 'navigation.gruvee.showMembersAction'
+
+export const SIDEMENU_ALL_MEMBERS_ID = 'navigation.gruvee.sideMenuAllMembers.id'
+export const SIDEMENU_ALL_MEMBERS_NAME = 'Members'
+
+export const USERNAME_INPUT_VIEW_ID = 'navigation.gruvee.usernameInputView.id'
+export const USERNAME_INPUT_VIEW_NAME = 'navigation.gruvee.usernameInputView'
