+++ app/config/navigation/themes.js
@@ -1,0 +1,10 @@
+import { BASE_BACKGROUND_COLOR } from '@StyleConstants'
+
+// Themes
+// eslint-disable-next-line import/prefer-default-export
+export const BASE_THEME = {
+    dark: true,
+    colors: {
+        background: BASE_BACKGROUND_COLOR,
+    },
+}
